extract animation for all menus into own css file
