+++ src/assets/styles/index.css
@@ -1,5 +1,6 @@
 @import '~@fortawesome/fontawesome-free/css/all.min.css';
 @import 'utilities/index.css';
+@import 'animations.css';
 
 @tailwind base;
 @tailwind components;
